backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
matplotlib==3.9.0
numpy==1.26.4
opencv-python==4.8.1.78
opencv-python-headless==4.8.0.74
pandas==2.2.2
pillow==10.3.0
scikit-image==0.23.2
torch==2.3.1
torchvision==0.18.1
yolov5==7.0.13
